Simplify handling of internal values a bit
The weak map that I previously talked about [1] wasn't as I intended it
to be, in this case the map only ever contain a single item.

My intention was to keep a map outside of the "test's runtime", as I am
doing with this change, but I now realize that strings aren't legal
WeakMap keys.

Furthremore, I like that the reference contained in the Cypress
environment is a primitive value. Less surprises this way.

Lastly, because the value known to Cypress is a primitive value and not
something where we're messing with method overloading, I am more
comfortable with reomving the opt-out feature of this.

[1] https://github.com/badeball/cypress-cucumber-preprocessor/issues/908#issuecomment-1367864681