Add a bunch of messages tests
These are essentially just duplication of the same scenarios used to
test JSON reporting. This didn't used to be necessary, because the JSON
report used to be a manifestation of messages.

I do however intent to remove use of `cucumber-json-formatter` and
generate these types of reports independently, thus I want independent
test coverage as well.